TabHost key handling corrected
When receiving key presses in TabHost.java, focus should be
requested for the selected tab indicator provided that the
following conditions are fulfilled:
1) A content view inside an embedded activity is currently focused.
2) No focusable view exists in the direction of the navigation
key press inside the embedded activity.
3) The TabWidget is located in the direction of the navigation
key press.
This should work for all locations of the TabWidget, not only
when the TabWidget is located above the tab content.

Change-Id: Ic5862cb72d2cc95aed9de92e42d6e0faf959fd43

Former-commit-id: 54ed87c685fe5a911a622ffef1d5824af6ea95a9