Fix crash downloading posters for HTML5 video.
We use the apache HTTP stack to download the poster images
for HTML5 video. This will crash if there is no host to use
as the "Host" header when making the request. Limit the java
poster downloader to just http or https for now. WebKit seems
able to display posters over other schemes like file://, but
doesn't always get the dimensions right. This fix just stops
us crashing.

Bug: 3180037
Change-Id: Idf51efda5b9ca1f2fe373c1fdb9c6bb7d5e254c8

Former-commit-id: 42509799f261c844bf6be66ce630c47bdee620f3