Fix bug with bad state in animators
Bug 3482310: The playing state was not being correctly set to
RUNNING after an animator was start()'d. Instead, we were seeking
to the start value (correct), setting the state to SEEKED (also correct),
but not resetting the playing state to STOPPED. So when the animation
actually started animating values, it didn't recognize that it was
starting a STOPPED animation, so it never set its state to RUNNING,
and never returned true from isRunning().

Change-Id: Iea92dce98f92f60052d8a9a451094b953f9f0c67

Former-commit-id: 154f14508a11627d5a995b6fe2a14a83d794a6fe