AIOOB exception fix in TabWidget
Bug http://code.google.com/p/android/issues/detail?id=15005

The problem was not specific to the legacy theme. The code that first
measure the tab's width with no contraint was incorrectly using the
mImposedTabsWidth array which could not have the right size if a
child was added.

The first measure after a child is added should indeed crash. Could
be investigated. This fix is sure anyway.

Change-Id: I4c0594e92019aabc545619e874bdfd49ce69c90e

Former-commit-id: cd59febcea2b75ebe657786133dcfa81fc3d5eb1