DO NOT MERGE Cherry-pick of 17fccc2086c [formerly cd59febcea2] from master
AIOOB exception fix in TabWidget

Bug http://code.google.com/p/android/issues/detail?id=15005

The problem was not specific to the legacy theme. The code that first
measure the tab's width with no contraint was incorrectly using the
mImposedTabsWidth array which could not have the right size if a
child was added.

The first measure after a child is added should indeed crash. Could
be investigated. This fix is sure anyway.

Change-Id: If5015aaa2d5574939fd5d6c6362ed6db94d35d4a

Former-commit-id: 4797f020cd255aa606d0bb1ca419de0d4facc3c6