Remove unnecessary SkImageDecoder_libjpeg.cpp
Adding files outside $(LOCAL_PATH) is a bad idea. It generates output files
to an unexpected directory. In this case, it's

out/target/external/skia/src/images/SkImageDecoder_libjpeg.o

It's not in $(PRODUCT_OUT), which is not correct.
Worse, it makes the build system be unable to build different
architectures (arm and x86) simultaneously, since the output files
conflcit.

Seems we can just remove SkImageDecoder_libjpeg.cpp since it's
already in libskia.

Change-Id: Ia0ed72d8a6cbd3b2f3997b71dfdcf72fee50f5bb

Former-commit-id: 4c1ac4e66560f53625de28663c67e81a5f8f4401