CheckBoxPreferences do not fire accessibility events
bug:4091160

We no longer allow not shown views to fire accessibility
events. Therefore, CheckBoxPreference can no longer ask its
associated checkbox to fire an accessibility event because
the latter is not attached to the view hierarchy
(i.e. not shown) by the time the CheckBoxPreference has a
chance to send an accessibility event. Instead the
CheckBoxPreference itself is responsible for firing the event.

Change-Id: Ia742882b13bf6c441cc76fa1e84b47a55cee601c

Former-commit-id: 1686e9afd4addf1655bf55b6d815e3f51934b654