b/4116658 Fix the jumpy scrolling with some flash site.
In the case that a touch event is passed to WebKit then back to
WebView, the coordinates will be converted from view to content
then back to view and the convertion could lose some accuracies.
For a flash content that only consumes TouchDown, all the
TouchMove events will go through this path and each time, the
data becomes more inaccurate. Even worse, the TouchMove event
updates the mScrollX/Y which will then be used for the convertion.
The effect amplifies really quick and the scrolling looks jumpy.

The fix is just to store the original view coordinates when pass
the event around.

Change-Id: Ie1424d7cfc6272348b194732e97168efe2dcf17b

Former-commit-id: 88b869a40a0eb386385a5de1f37585458c65b548