Airplane mode on always timed out when camping on LTE
While on LTE, GsmDataConnectionTracker is implicitly associated with CdmaServiceTracker.
CdmaServiceTracker asynchronously waits for the disconnection of data before turning off
the Radio but GsmDataConnection does not notify the disconnection due to different logic
in GsmServiceTracker.

This fix is to have common airplane mode toggle logic on both GSM and CDMA flavors.

Change-Id: I67ff82a59a2243856ae3e7c5e37b5726bcbf3c94

Former-commit-id: 95bc2b7564b201d10f2a5e319f69972e18c58261