stagefright aacenc/amrwbenc: Remove trailing whitespace
This was applied in AOSP master earlier, in commit
b21611e5b0f1797e39e781d91ade023bac87637f, but when merged
into the Google internal tree in
2bd8a0102eea8a13a2af45e8bcd8246a8c1240ae [formerly 88c2d4d26d79384f32a25bd775cf873cb678b12a], none of these
changes have been brought in. (The diff of this merge commit
and its first parent, d94b71d, is empty.)

Therefore, reapply the rest of these changes on the files
that still are present in the current version.

Change-Id: I645a9b9134769958a11808a06163123d28d83e46

Former-commit-id: 62a62ef415bd1db5f5862e5ef1cc127ccd6465b8