Add RIL_UNSOL_RIL_CONNECTED and pass radio technology + 2 to setupDataCall.
Added CommandsInterface.registerForRilConnected which allows components
to register for when the ril connects and provides the version of the
ril to the registrants or a -1 if the ril disconnects.

For ril version 6 RIL.setupDataCall needs to have RadioTechnology + 2
as the radioTechnology parameter when the rilVersion is >= 6.

DataConnection then uses registersForRilConnected to get the ril
version and provide the proper radioTechnology value to
RIL.setupDataCall

Change-Id: Iacba764b9e38a4a138d186ccf1625fea760cc56d

Former-commit-id: 1dd2ada5cf7d0c95afb4731d082473f34ee7f433