Increment database version to 65
The onUpgrade path was upgrading the database version to 65, but the
current version was marked as 64. That led to the database being
upgraded to 65 and then wiped because it didn't match 64.

This was introduced in HC change 37405725a0b4eabbcc1544c0ec962e6de69864b4 [formerly 54d068ec6af0ee6d261a135400efe6816c6f5ffe]

Bug: 4319406
Change-Id: Ib6efcf34e820948d23d3a2b8ef3afc9012a93c22

Former-commit-id: aef35dde4d687b98146c449639a06f21228a0acc