Handle DHCP renewal and wakeup in framework
The native DHCP client renews IP address, but can fail to do so when the device
is in suspend. This has lead to issues where device fails to renew in time and
on networks with lots of devices re-using expired DHCP IP addresses, it leads to
severe issues.

Handle DHCP renewal and wakeup in framework

Bug: 3344732
Change-Id: Ie4062e04a477f4a233946155e40a7c999b337c3f

Former-commit-id: d2127c438fc858a4f6c8ad75b7a1e4491da69a78