All SuggestionSpans are removed when the window loses focus.
SuggestionSpans should have a limited, well defined life-span.
Saving these in onSaveInstanceState is inappropriate since these can
be very large and will not necessarily be saved/loaded by the application.

Keeping them in the application when it is dismissed is not consistent
since they may be garbage collected while the application is not active.

Change-Id: Icebf6cb0218528f9d162a689ad33efeb770fbede

Former-commit-id: ba4997ac355d11db3c7b04d4215c6948d05cb047