TelephonyRegistry remove 4 uses of FLAG_RECEIVER_REPLACE_PENDING.
The use of FLAG_RECEIVER_REPLACE_PENDING is an optimization to reduce
redundant messages. For some calls this is fine, such as
broadcastSignalStrengthChanged. For others it can cause problems, such
as broadcastDataConnectionStateChanged where the MobileDataStateTracker
will not be able to properly track the state if messages are dropped.
This changes removes the optimization from all methods in
TelephonyRegistry except broadcastSignalStrengthChanged.

bug: 4427303
Change-Id: Ia9c0dd0ce66b8b3dcda770f5ab2c63cf08f3ebfc

Former-commit-id: 71f6cb1324ecec9206c6bde1ab8abd31199b1927