change how test sets timeout for webkit dump
there are cases where webkit dump timed out, but a RuntimeException
was not thrown (as it should have) after the timeout limit was
reached. It's probably because the Looper message to invoke time
out is obtained after the message to dump webview. This change
adjusts the order of obtaining the message so hopefully it can make
the timeout work for real.

Change-Id: I4197699f5aad765c88dc81123d052e4c49d7c53d

Former-commit-id: 97499df874f680eaa55d96074f3dcca34685ca12