stagefright amrwb: Make local const arrays static
Some of these tables exist in both the amrnb and amrwb libraries,
but with different content.

This avoids some cases of multiple definitions of symbols, if
linking in both libraries statically into the same executable,
if the libraries are built as C. (As C++, the local symbols
get mangled so they don't conflict with the external C symbols
in amrnb, that are unmangled.)

Most of this was committed on the master branch of the opencore
repository in commit 0cd4b3ac412dd0a8370ce339e89cf346c6cfe395,
based on AOSP contribution 10016.

Change-Id: I5387e699ce54a56dc43fcfa0396dee5c99280fd9

Former-commit-id: b3094e4a0ea10c31643acd6422c2b214d4c14f80