stagefright amrnb: Fix a bug on architectures where long is 64 bit
Initially, this code gave a warning "comparison is always
false due to limited range of data type" if long was 64 bit.
Additionally, the output from the AMR encoder was mostly noise.

This was committed on the master branch of the opencore
repository in commit f532d145194e474cb30d8644b8eee83873ad24cc,
based on AOSP contribution 10904.

Change-Id: I822dcc74beeae548d3032ffda759de9c8cd1fcfe

Former-commit-id: 69c2edfaf71f5494ab7a3caf26e8ab782d49781d