Fix overview scale setting for mobile sites.
This fix will only set the overview scale in case the screen
is rotated from landscape to protrait, since in this case
the overview scale will be smaller and need explicit setting.
For the other way around, there's no need to set and it has wrong
effect for mobile sites.

issue: 4343683
Change-Id: I92cbf848bc2ed4184bd0c6b67992ff5cbc633c9f

Former-commit-id: 73bb5734f06adc3b61a102fcb2f5295caa9e4886