Use unchecked admin lookup for getStorageEncryption()
Due to a copy/paste typo, getStorageEncryption() was protected by
getActiveAdminForCallerLocked(), which made the API unreadable if the
caller is not an admin.  This doesn't match the rest of the DPM API which
allows the "get" calls to be made anonymously.

This fix supercedes the proposed workaround CL in the DPM ApiDemo sample.

Change-Id: If1a6237634e97ced09a7c6a8876bb6b0f60c9be9

Former-commit-id: c994d693b684bcba1a690310b5b2be306074cc02