Take advantage of EGL_BUFFER_PRESERVED on IMG.
The IMG architecture does not support the EGL_SWAP_BEHAVIOR_PRESERVED_BIT
in EGL configurations. This bit is used only to be able to change the
EGL_SWAP_BEHAVIOR, which matters when the default behavior is not the
one we want. However on IMG, the default swap behavior is the one we
desire.

Change-Id: Iff5980c97df9e41bc837e424baf83b8eadbb7161

Former-commit-id: 6f7d9394ec69e9fb38ca5fc2caf6d2aef6f7442d