DO NOT MERGE Use unchecked admin lookup for getStorageEncryption()
Due to a copy/paste typo, getStorageEncryption() was protected by
getActiveAdminForCallerLocked(), which made the API unreadable if the
caller is not an admin.  This doesn't match the rest of the DPM API which
allows the "get" calls to be made anonymously.

This fix supercedes the proposed workaround CL in the DPM ApiDemo sample.

Backport of: If1a6237634e97ced09a7c6a8876bb6b0f60c9be9

Change-Id: I728b2d85878b7c73a063c906529f0d973e166b49

Former-commit-id: 237574444b311f2f4d23c62f63162f77d34f10bc