Fix issue #4539687: At least one compatibility-mode app (Androminion)...
...will only launch when held in portrait mode.

There was a bug in the window manager that caused all of the careful code to
update the configuration in sync with movements between activities to break.
Now it is fixed, so this app works, and we no longer see the bad slow orientation
changes when switching between activities that want to be in different
orientations.

Change-Id: I5d93f99649849bdaca2e8bebade6b91b8b6cf645

Former-commit-id: bc7386c2615a7c496deea59ac1b5ab2da7f7b61f