Always clear the Message when recycled.
Clear the Message when it is recycled, even if it will not go
back into the Message pool.  This makes the behavior of recycle()
more consistent and ensures that the Message does not hold onto
other object after it has been recycled (useful in case there are
stale references to the Message lingering elsewhere).

Change-Id: I26b6a4b629f9c0b6bed70fdc42734919f30e64c4

Former-commit-id: fc9ff4c834c5b3a3c4b024b89375d6c329092791