Ensure the RemoteViewsAdapter cleans up its HandlerThread.
Using a finalizer for this is kind of gross but it is robust.
We could try to hook onDetachedFromWindow() on the View
but that would introduce additional lifecycle management to
the adapter since a View could potentially be detached and then
reattached, so we might need to recreate the handler, etc.

Ideally this code should probably be changed to use AsyncTask.

Change-Id: I9ad5fff388411619363e4a075ed0244966b38726

Former-commit-id: fc442bdea14289656ef1f537103578eb71faf473