NativeDaemonConnector: Fix SDCard mounting bug in the emulator.
This patch fixes a really annoying bug that prevented the emulator
from properly mounting the SD Card.

The issue was that mOutputStream was not initialized when the onDaemonConnected
callbacks are called, just after the daemon's socket connection is established.

In one specific case (MountService a.k.a. VoldConnector), the corresponding
callback function immediately tried to perform a doCommandList() call on the
connector object. The request failed because mOutputStream was still defined
as null when it happened.

By moving the initialization of the variable just before the callback is
invoked, we ensure that things work properly.

Change-Id: I46726a50ef69eaafe70081ff4eda1244b744beba

Former-commit-id: 2435a927f30efe38c673bfaed64a9880028da9e8