Contacts contract changes for social updates.
This change adds support for a richer API supporting social status
updates.

The new API supports providing multiple status updates (stream items)
to be associated with a single raw contact, and for multiple photos
(stream item photos) to be associated with each stream item.  A
limited number of stream items will be stored - callers can query for
the limit to optimize calls.  Any inserts beyond the limit will
result in expiration of the oldest item.

Change-Id: Ib46fdf8864472d4a1be250b9051e7d37ec32c089

Former-commit-id: d414f108e70766b1a9d6f4e99201f960cfa45310