Data connection retry pattern could be broken by PDP sharing
If a DataConnection is pending re-connect alarm, the new request
from another ApnContext sharing the same DC could disrupt
the re-connection pattern currently engaged.

This patch is to ensure the new request for PDP sharing
scenario will not trigger another SETUP_DATA request if
reconnection alarm is set in AlarmManager.

Bug: 4901019
Change-Id: I98b0d9af8b58fb880efdbc0246009de5cb116a54

Former-commit-id: 1e5c72c7041e0c8299cb5a5d028c022eb01036c6