CallerInfo query: Turn on "geo description" lookup for unknown numbers
With the latest version of the external/libphonenumber library
<https://android-git.corp.google.com/g/117190> we now have real data for
the PhoneNumberOfflineGeocoder.getDescriptionForNumber() feature.

So enable it in the CallerInfo query.  This means that the incoming call
UI will now show a geo description like "California" or "Maryland" (along
with the number) for unknown incoming numbers.

I also needed to work around an issue with the latest phonenumber library:
The new library apparently requires countryIso to be uppercase (e.g. "US")
but the CountryDetector.detectCountry().getCountryIso() call currently
returns "us", which causes PhoneNumberUtil.parse() to fail.  (This also
broke some ContactsProvider tests too.)  So force the countryIso to
uppercase for now.

TESTED (on Crespo):
  - Incoming calls from an unknown number:
    ==> State name is now displayed.
  - Incoming calls from a number that matches a contact:
    ==> no change in behavior

Bug: 4595580
Change-Id: I03126e1eee99d428e76bbbad5b3856be3874f549

Former-commit-id: cec25c4e8afdc56451f7405f8605c1d67433e2ff