Introduced 'SELF_CHANGE' extra for voicemail notifications.
A follow up change in voicemail content provider will now
populate this extra instead instead of EXTRA_CHANGED_BY to indicate if
the change was caused by the receiver. This makes the the security of
the notifications a bit tighter by not revealing the package name of the
app that made the change to others.

See Bug http://b/4773134

Change-Id: Ie376c40c7bba9792da3b18e66d9f205166a402e0

Former-commit-id: 4687cbc48d710069d65086c7e8b600edb01e1255