Fix 4689527: Update Recents
Fixed layout in landscape mode where recents only filled half the screen
because "match_parent" in the layout was being ignored. The fix is to
have a temporary ViewGroup as the parent to ensure the layout parameters
are valid.

Changed to solid 70% opaque background on phone.

Removed "dismiss" chevron from phone layout.

Removes glow on phones.

Change-Id: Id5319eeba4fdd8c9ef5792168b35162cc1bbfe38

Former-commit-id: f2a16b27afd1ede5880d05eb9e19090c7e3e820e