Add the "debug.jni.logging" property so third-party developers can debug their JNI problems.
This turns on the super-verbose but indispensible logging of all native method
calls and all calls to JNI functions (for third-party code only). In particular,
if you have a local reference bug, you can search for the reference given in
the crash and see exactly where it came from. In every case I've seen so far,
that's pinpointed the bug exactly.

Change-Id: Ifb7ba02ae637bdd53cd8500febdcb9d4d7799bda

Former-commit-id: ae07ecf3766c38af1c12822458b98036b28bd4c0