Always redraw the UI on orientation change Bug #4994715
In some cases, on some architectures, the hardware renderer could fail
to properly render a window after an orientation change. The fix is
to force the GL driver to drop any existing buffer to make sure the
next buffer used to draw has the correct geometry.

This change also fixes an issue that was preventing partial screen
updates from being properly enabled.

Change-Id: I874a48590845a8624425130af7db8e20022895dd

Former-commit-id: 7e1160e8664a9c5c79367f4ba3c2266a39bbcef5