Bug 4534376: Suggestion popup is not clipped.
This happened when the popup was larger than the screen because of a very long suggestion.
The text will now be ellipsized with a marquee (better than using 2 lines I believe).

Also removed the up layout that was intended to be used when the popup did not fit vertically.
The popup is moved up instead if needed.

Change-Id: Ie741610e74ade10c691614f1c30d0e17b213c698

Former-commit-id: 31a41d35f64c016c63eebaaceff938a0ca40b122