Fix leak in LayoutTransition
Static objects were referencing the first LayoutTransition object,
which referenced its target container, which eventually referenced the
Activity. That reference chain never went away.
The fix is to create animators that do not reference any target
object (they are just templates which are loaded with real target
objects when they are started).

Change-Id: Ie29f53bf3b886b8052b6ee3a56647a312d9adeaa

Former-commit-id: 622e05c4d2f7c7542ab9cdd30c31813cb6cdb103