Replace DroidKufi-Regular with DroidNaskh-Regular
Naskh is a fancier font than Kufi and better for use with web page
content. Until we have more refined support for fallback fonts, only
one can be available at a time. This change switches from Kufi for
the purpose of comparitive testing to see which should be chosen
for ICS.

Change-Id: I081a002b6f72461f7ce18a86b3dec36c7f0f7506

Former-commit-id: 312cae7d39fdcadf36d9a0b1cace61f068859c3c