Change ZipFile flags anonymous enum
The arguments that ZipFile takes were apparently meant to be a typed
enum, but the name of the type was taken out because it creates invalid
conversion warnings in C++ since this is an invalid use of enum.

Just change the typedef enum to an anonymous enum so we get rid of the
compiler warnings. It will be implicitly converted to int instead.

Change-Id: Ie72c4c235c8f9ee8fd1b795e1010b77c55e983de

Former-commit-id: af2872d1e324ed67b44b4f44e2d83b1f0d985027