Added ACCOUNT_TYPE_AND_DATA_SET pseudo-column.
This was added to simplify some logic in the provider and app, as
there are places that can simply compare the concatenation of the
account type and data set.

Also added a DATA_SET extra for contact creation, to allow for this
value to be passed around via intents.

Bug 5077096

Change-Id: Ia0634cfcf4590fd90ccb5cb99360c38c86f08d25

Former-commit-id: 1d55c3343d6eb89176fb057233942887466c1b68