Reduce memory usage of status bar expanded view.
It was configuring this to have its width follow the display size
but height be a fixed amount.  As a result, during a rotation we
would end up with a surface that is 800x712, which uses a lot more
memory.

The fix is to just always set the window to a fixed size, changing
it when the display size changes.

Also the expanded view was setting itself to use a hardware layer
for no reason -- it is a top view so there is no point in this,
and anyway it is doesn't even use hw rendering.  This saves 1.5MB
of the layer bitmap.

This change also fixes the returned problem where the expanded
view would flicker when pulling it down in landscape.

Change-Id: If57420b0bc3fdc2706d2d3b36cb2d287b5fc9e27

Former-commit-id: 1dacf2772d0099ae74f42f81d162a3d0e180ffb7