Fix bug 5135608 - Let AbsListView take focus while detached
While it is valid to request focus while detached from a window,
AbsListView did not handle it gracefully if it had an adapter whose
data changed while the view was detached, since DataSetObservers are
unregistered during this time. Assume that the data could have changed
if we gain focus while detached.

Change-Id: Id135693ffcc48ff5a9d530d335ab449180624c43

Former-commit-id: b3750133db256ffa5956c8836bc86c3c22918cb1